UniqueIDs from Runs are now migrated in the background.
Unique IDs are migrated in the background, however if someone attempts to
retrieve the unique id before it haen migrated then it will be migrated on
demand by the loading of the Run.

The code now requires JDK7 to run in order to handle creation of IDs in a
thread safe manner.